Modify hash_file.py and upload testScript.py
hash_file.py has been modified to allow for execution from the command
line with two additional args following the script name.  Arg1 is the
path to the file as a string.  Arg2 is the hashing algorithm as a
string.

testScript.py can be used to check script execution from command line.